requirements.txt
mcp>=1.2.0
openai>=1.0.0
qdrant-client>=1.6.0
python-dotenv>=1.0.0 